Move peerinfo and localinfo from tcp to socket
With the addition of UDP sessions, this information needs to be
available to both L4 consumers. Technically speaking, socket.rb
can present an L3-only construct (raw IP sockets), which wont have
port information. However, this has not been an issue when using
the new transports in testing.

This is a breakout of the UDP handlers and payloads PR (#6035).