Implement a certificate provider pattern in Socket
Rex utilizes a class variable in the ThreadFactory which permits
passing a thread factory down from consumers so long as it matches
the API. The benefit of such an arrangement is that the consumer
is responsible for the dependencies of the provider object, which
in turn lets Rex keep its deps to a minimum.

Following that pattern, Rex::Socket::Ssl gets a @@cert_provider
variable which by default uses the existing module methods now
attached to a CertProvider namespace in order to create certs and
attributes.

Msf can now designate its own provider at runtime in order to give
users greater control, entropy, etc, without weighing down this
gem or obsoleting useful examples of how Rex::Text can be utilized
by consumers.

Also added minor cleanup of prior commit replacing instance method
calls to the module methods referencing SslTcpServer with just the
Ssl namespace (which in turn proxies calls to the @@cert_provider).

Testing:
  None, this is part of a larger effort spanning a couple of repos.