Address Brent's comment - drop @@loaded_openssl
MRI is compiled with OpenSSL by default, so the graceful failure
to load OpenSSL is basically a dead codepath.

Remove the openssl loading and checking blocks, and require it
as a normal gem.

Note: this might create problems with other interpreters, so we
might need to revisit this approach as a way for them to load the
runtime-specific ssl gems down the road.