Prevent DNS leaks & pivot using optional resolver
`Rex::Socket` eventually ends up relying on the standard library
for name resolution in socket-related calls. Specifically, the use
of `Socket.gethostbyname` and `Addrinfo.getaddrinfo` can lead to
the use of system DNS, recursing into places where users may not
want to send DNS requests.

Implement a mechanism to set a global name resolver for all of
`Rex::Socket` producing `.get('msf.io')` output analogous to that
of `Rex::Proto::DNS::Resolver`. Once that resolver is in-place,
the class method calls will be diverted to internal ones using
said resolver to create API-compatible output structures
feeding transparently back to the caller. Unlike approaches using
`Rex::Socket::SwitchBoard` directly, this mechanism can't create
loops in the stack as the `Resolver` uses IP-based NS' and cannot
get stuck looking up the name servers against which to look-up...

Clean up raw calls to `Socket` and ensure that all `gethostbyname`
calls are properly wrapped.

Note: a fully-compatible resolver would cache DNS requests until
the TTL on them expires and honor the local system hostsfile just
like the native standard library does. However, that would fall
into `rex/proto` or other parts of the ecosystem and not in-scope
for this repository. As a result, the operating semantic of this
library is unchanged by default and requires input of a relevant
`Resolver` from other components which are capable of constructing
and configuring one as appropriate.

Resolver tasks:
  Optimize for performance and concurrent access (w/ Rex::Proto)
  Implement record caching layer similar to (or reusing from)
`Rex::Proto::DNS::Server`
  Implement hostsfile mechanism (descendant class from the base
`Resolver` or via mixin/extend semantic)
  Review options for per-caller instantiation or NS passing by the
caller to permit dynamic dispatch to pivoted nameservers past the
`SwitchBoard` - this is likely to be handled by the mechanism which
will deal with the hostsfile concern such as to use session and
workspace data to determine which NS to query for an FQDN and how.